apps/LUBM/pyproject.toml
[project]
name = "lubm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Vincent Van Schependom", email = "vincent.van.schependom@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "omegaconf>=2.3.0",
    "owlrl>=7.1.4",
    "rdflib>=7.5.0",
]

[project.scripts]
lubm = "lubm:main"

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"
